squidpro-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyjwt==2.8.0
httpx==0.27.0
python-dotenv==1.0.1
